remove wrapper component
