--- packages/components/molecules/f-restaurant-card/src/assets/vue/mixins/errorBoundary.mixin.js
@@ -1,11 +1,0 @@
-import RenderlessSlotWrapper from '../../../components/RenderlessSlotWrapper';
-
-export default {
-    props: {
-        // An optional safeguard component to wrap around restaurant card data points such as ratings, cuisines etc.
-        errorBoundary: {
-            type: Object,
-            default: () => RenderlessSlotWrapper // by default returns a renderless component that will just render it's slot and not bloat markup
-        }
-    }
-};
