--- packages/components/molecules/f-restaurant-card/src/components/restaurantCardVersions.js
@@ -1,7 +1,0 @@
-import RestaurantCardV1 from './RestaurantCard.v1.vue';
-
-const components = {
-    v1: RestaurantCardV1
-};
-
-export default { components };
