+++ packages/components/molecules/f-restaurant-card/src/constants/custom-events.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
+export const EVENT_CLICK_RESTAURANT_CARD = 'click-restaurant-card';
